// core-web/libs/sdk/react/src/lib/next/types.ts
export interface DotPageAssetLayoutRow {
    identifier: number;
    value?: string;
    id?: string;
    columns: DotPageAssetLayoutColumn[];
    styleClass?: string;
}

export interface DotCMSVanityUrl {
    pattern: string;
    vanityUrlId: string;
    url: string;
    siteId: string;
    languageId: number;
    forwardTo: string;
    response: number;
    order: number;
    temporaryRedirect: boolean;
    permanentRedirect: boolean;
    forward: boolean;
}

export interface DotPageAssetLayoutColumn {
    preview: boolean;
    containers: DotCMSColumnContainer[];
    widthPercent: number;
    width: number;
    leftOffset: number;
    left: number;
    styleClass?: string;
}

export interface DotCMSColumnContainer {
    identifier: string;
    uuid: string;
    historyUUIDs: string[];
}